Cleanup
- Broke out logic
- Deleted stuff no longer needed from filter
- Deleted other stuff from operation
